requirements.txt
streamlit
langchain
langchain-community
langchain-openai
pypdf
python-dotenv
faiss-cpu

